+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 3.2.0
 ====
 
+* The `summary` method on articles is now HTML-aware, and can be provided with
+  a different summary length or ellipsis string: `summary(1000, '&hellip;')`
 * The `summary_generator` option now recieves the rendered article (without
   layout) instead of the template source.
 * Set `summary_length` to false to always use the full article as a summary.